rewrite.jsをrewriteした
