+++ html/js/assets/stage_glyph.json.js
@@ -1,0 +1,1 @@
+window.gLocalAssetContainer["stage_glyph"] = "%7B%0A%09%22map%22%3A%7B%0A%09%09%2248%22%3A%7B%0A%09%09%09%22x%22%3A0%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2249%22%3A%7B%0A%09%09%09%22x%22%3A50%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2250%22%3A%7B%0A%09%09%09%22x%22%3A100%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2251%22%3A%7B%0A%09%09%09%22x%22%3A150%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2252%22%3A%7B%0A%09%09%09%22x%22%3A200%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2253%22%3A%7B%0A%09%09%09%22x%22%3A250%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2254%22%3A%7B%0A%09%09%09%22x%22%3A300%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2255%22%3A%7B%0A%09%09%09%22x%22%3A350%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2256%22%3A%7B%0A%09%09%09%22x%22%3A400%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%2C%0A%09%09%2257%22%3A%7B%0A%09%09%09%22x%22%3A450%2C%0A%09%09%09%22y%22%3A0%0A%09%09%7D%0A%09%7D%2C%0A%09%22missingGlyph%22%3A%7B%0A%09%09%22x%22%3A500%2C%0A%09%09%22y%22%3A0%0A%09%7D%2C%0A%09%22width%22%3A50%2C%0A%09%22height%22%3A62%0A%7D"
