+++ html/js/game.json.js
@@ -1,0 +1,1 @@
+window.gLocalAssetContainer["game.json"] = "%7B%0A%09%22width%22%3A%20790%2C%0A%09%22height%22%3A%20600%2C%0A%09%22fps%22%3A%2030%2C%0A%09%22main%22%3A%20%22.%2Fscript%2Fmain.js%22%2C%0A%09%22assets%22%3A%20%7B%0A%09%09%22main%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2Fmain.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22Spinner%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FSpinner.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22BorderRect%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FBorderRect.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22MineSweeper%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FMineSweeper.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22GameScene%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FGameScene.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22Cell%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FCell.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22ResultScene%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FResultScene.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22TitleScene%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FTitleScene.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22glyph%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22text%22%2C%0A%09%09%09%22path%22%3A%20%22text%2Fglyph.json%22%0A%09%09%7D%2C%0A%09%09%22stage_glyph%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22text%22%2C%0A%09%09%09%22path%22%3A%20%22text%2Fstage_glyph.json%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22triangle%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Ftriangle.png%22%2C%0A%09%09%09%22width%22%3A%2045%2C%0A%09%09%09%22height%22%3A%2042%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22triangle_on%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Ftriangle_on.png%22%2C%0A%09%09%09%22width%22%3A%2045%2C%0A%09%09%09%22height%22%3A%2042%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22triangle_reverse%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Ftriangle_reverse.png%22%2C%0A%09%09%09%22width%22%3A%2045%2C%0A%09%09%09%22height%22%3A%2042%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22triangle_reverse_on%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Ftriangle_reverse_on.png%22%2C%0A%09%09%09%22width%22%3A%2045%2C%0A%09%09%09%22height%22%3A%2042%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22number%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fnumber.png%22%2C%0A%09%09%09%22width%22%3A%20440%2C%0A%09%09%09%22height%22%3A%2034%0A%09%09%7D%2C%0A%09%09%22stage_number%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fstage_number.png%22%2C%0A%09%09%09%22width%22%3A%20550%2C%0A%09%09%09%22height%22%3A%2062%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22start%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fstart.png%22%2C%0A%09%09%09%22width%22%3A%20330%2C%0A%09%09%09%22height%22%3A%20100%0A%09%09%7D%2C%0A%09%09%22start_on%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fstart_on.png%22%2C%0A%09%09%09%22width%22%3A%20330%2C%0A%09%09%09%22height%22%3A%20100%0A%09%09%7D%2C%0A%09%09%22title%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Ftitle.png%22%2C%0A%09%09%09%22width%22%3A%20466%2C%0A%09%09%09%22height%22%3A%20320%0A%09%09%7D%2C%0A%09%09%22clear%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fclear.png%22%2C%0A%09%09%09%22width%22%3A%20590%2C%0A%09%09%09%22height%22%3A%20390%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22gameover%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fgameover.png%22%2C%0A%09%09%09%22width%22%3A%20590%2C%0A%09%09%09%22height%22%3A%20390%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%2C%0A%09%09%22open_cell%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fopen_cell.png%22%2C%0A%09%09%09%22width%22%3A%2080%2C%0A%09%09%09%22height%22%3A%2060%0A%09%09%7D%2C%0A%09%09%22close_cell%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22image%22%2C%0A%09%09%09%22path%22%3A%20%22image%2Fclose_cell.png%22%2C%0A%09%09%09%22width%22%3A%2080%2C%0A%09%09%09%22height%22%3A%2060%0A%09%09%7D%2C%0A%09%09%22Button%22%3A%20%7B%0A%09%09%09%22type%22%3A%20%22script%22%2C%0A%09%09%09%22path%22%3A%20%22script%2FButton.js%22%2C%0A%09%09%09%22global%22%3A%20true%0A%09%09%7D%0A%09%7D%0A%7D"
