+++ html/js/initGlobals.js
@@ -1,0 +1,5 @@
+window.g = require("@akashic/akashic-engine");
+
+if (! ("gLocalAssetContainer" in window)) {
+	window.gLocalAssetContainer = {};
+}
